-u swagger, auth.routes, views, test dll. & add nodemailer, websocket, sentry dll.
